fix db schema imports
